backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0
pandas>=2.2.3
numpy>=2.1.0
scikit-learn>=1.5.2
PyJWT>=2.9.0
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-asyncio>=0.22.0
httpx>=0.24.0
anyio>=3.7.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
sentence-transformers>=2.2.2
torch>=2.1.0
aiosqlite>=0.18.0

